+++ LittleEngine/LittleEngine/ModuleCamera.cpp
@@ -84,7 +84,7 @@ void ModuleCamera::EnableOrthographic() {
 }
 
 void ModuleCamera::LoadProjection() {
-	aspect = Engine->moduleWindow->width / Engine->moduleWindow->height;
+	aspect = static_cast<float>(Engine->moduleWindow->width / Engine->moduleWindow->height);
 	if (perspectiveEnable) {
 		projection = glm::perspective(glm::radians(frustumFov), aspect, nearPlane, farPlane);
 	}
