+++ LittleEngine/LittleEngine/ModuleWindow.cpp
@@ -67,8 +67,8 @@ bool ModuleWindow::CleanUp()
 
 void ModuleWindow::WindowResized(unsigned width, unsigned height)
 {
-	this->width = width;
-	this->height = height;
+	this->width = static_cast<int>(width);
+	this->height = static_cast<int>(height);
 	Engine->moduleCamera->LoadProjection();
 	SDL_SetWindowSize(window, width, height);
 	glViewport(0, 0, width, height);
